+++ apps/frontend/localization/content/en/home.md
@@ -1,0 +1,11 @@
+# Welcome to HumanDBS
+
+## Static Markdown content
+
+Markdown が動的にimportされた`*.md`ファイルを表示するためのコンポーネントです。
+`*.md`ファイルは、`src/content` ディレクトリに保存されています。
+
+## Static Markdown content
+
+GitHub に保存されたMarkdownファイルを表示することもできる。
+[/contact](/contact) にアクセスすると、公的のGitHub レポ（tailwindcss のリポ）に保存された `readme.md` が表示される。
