+++ apps/frontend/localization/content/ja/about.md
@@ -1,0 +1,69 @@
+# 簡単なMarkdown ファイル
+
+This is a sample markdown file. You can use markdown to write your content.
+
+## Heading 2
+
+### Heading 3
+
+#### Heading 4
+
+##### Heading 5
+
+###### Heading 6
+
+**Bold text**
+
+_Italic text_
+
+- List item 1
+- List item 2
+  - List item 3
+    - List item 4
+
+1. Numbered list item 1
+2. Numbered list item 2
+   1. Numbered list item 2.1
+      1. Numbered list item 2.1.1
+      2. Numbered list item 2.1.2
+
+[Link to Google](https://www.google.com)
+
+![Image](https://via.placeholder.com/150)
+
+```javascript
+console.log("Hello, world!");
+```
+
+> Some blockquote text
+> Some more blockquote text
+
+---
+
+Horizontal rule
+
+---
+
+| Header 1 | Header 2 | Header 3 |
+| -------- | -------- | -------- |
+| Row 1    | Row 1    | Row 1    |
+| Row 2    | Row 2    | Row 2    |
+| Row 3    | Row 3    | Row 3    |
+
+---
+
+- [x] Task 1
+- [ ] Task 2
+
+---
+
+<details>
+  <summary>Click to expand</summary>
+
+Some hidden content
+
+</details>
+
+---
+
+<iframe width="560" height="315" src="https://www.youtube.com/embed/dQw4w9WgXcQ" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
