+++ apps/frontend/localization/content/ja/home.md
