+++ apps/frontend/messages/en.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://inlang.com/schema/inlang-message-format",
+  "example_message": "Hello {username}"
+}
