+++ apps/frontend/messages/ja.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "https://inlang.com/schema/inlang-message-format",
+  "example_message": "{username}様、こんにちは！"
+}
