+++ apps/frontend/package.json
@@ -9,13 +9,14 @@
     "start": "bun ./output/server/index.mjs"
   },
   "dependencies": {
+    "@formatjs/intl-localematcher": "^0.6.1",
     "@radix-ui/react-slot": "^1.1.1",
     "@tailwindcss/postcss": "^4.1.10",
     "@tailwindcss/vite": "^4.0.3",
     "@tanstack/react-query": "^5.62.16",
-    "@tanstack/react-router": "^1.121.2",
-    "@tanstack/react-router-with-query": "^1.121.2",
-    "@tanstack/react-start": "^1.121.7",
+    "@tanstack/react-router": "^1.121.34",
+    "@tanstack/react-router-with-query": "^1.121.34",
+    "@tanstack/react-start": "^1.121.35",
     "@tanstack/react-table": "^8.20.6",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -23,6 +24,7 @@
     "i18next-browser-languagedetector": "^8.2.0",
     "i18next-fs-backend": "^2.6.0",
     "lucide-react": "^0.474.0",
+    "negotiator": "^1.0.0",
     "postcss": "^8.5.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
@@ -38,12 +40,15 @@
   "devDependencies": {
     "@humandbs/backend": "workspace:*",
     "@humandbs/eslint-config": "workspace:*",
+    "@inlang/paraglide-js": "2.1.0",
+    "@inlang/paraglide-vite": "^1.4.0",
     "@tailwindcss/typography": "^0.5.16",
     "@tanstack/eslint-plugin-query": "^5.66.0",
     "@tanstack/eslint-plugin-router": "^1.114.12",
     "@tanstack/react-query-devtools": "^5.80.7",
     "@tanstack/react-router-devtools": "^1.120.16",
     "@tanstack/router-plugin": "^1.114.25",
+    "@types/negotiator": "^0.6.4",
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.6",
     "@vitejs/plugin-react": "^4.5.1",
