+++ apps/frontend/project.inlang/.gitignore
@@ -1,0 +1,1 @@
+cache
