+++ apps/frontend/project.inlang/project_id
@@ -1,0 +1,1 @@
+kdCAHSXxpmznNHCIoo
