+++ apps/frontend/src/components/LanguageSwitcher.tsx
@@ -1,31 +1,52 @@
-import { useState } from "react"
+import { cn } from "@/lib/utils";
+import { i18n, Locale } from "@/serverFunctions/i18n-config";
+import {
+  useNavigate,
+  useRouteContext,
+  useRouter,
+} from "@tanstack/react-router";
 
-import { cn } from "@/lib/utils"
-
-const languages = [
-  { code: "en", name: "En" },
-  { code: "ja", name: "Ja" },
-] as const
+import { useState } from "react";
 
 export function LangSwitcher() {
+  const { navigate, invalidate } = useRouter();
 
-  const [selectedLang, setSelectedLang] = useState<string>(() => languages[0].code)
+  const { lang: currentLang } = useRouteContext({ from: "/$lang" });
+
+  const [count, setCount] = useState(0);
+
+  function handleSwitch(lang: Locale) {
+    navigate({ to: ".", params: { lang } });
+    invalidate();
+  }
 
   return (
-    <div className=" bg-primary/50 f relative flex rounded-full p-1">
-      {
-        languages.map((lang) => (
-          <button onClick={() => {
-            setSelectedLang(lang.code)
-          }} key={lang.code} className={cn(" z-10 uppercase font-bold cursor-pointer rounded-full h-8 w-8 text-center text-foreground-light text-sm", {
-            " text-white": selectedLang === lang.code,
-          })}>{lang.name}</button>
-        ))
-      }
+    <div className="bg-primary/50 f relative flex rounded-full p-1">
+      {i18n.locales.map((lang) => (
+        <button
+          onClick={() => {
+            handleSwitch(lang);
+          }}
+          key={lang}
+          className={cn(
+            "text-foreground-light z-10 h-8 w-8 cursor-pointer rounded-full text-center text-sm font-bold uppercase",
+            {
+              "text-white": currentLang === lang,
+            }
+          )}
+        >
+          {lang}
+        </button>
+      ))}
+      {count}
+      <button onClick={() => setCount((p) => p + 1)}> + </button>
 
-      <div className=" bg-secondary absolute z-0 size-8 rounded-full transition-transform" style={{
-        transform: `translateX(${selectedLang === "en" ? 0 : 32}px)`,
-      }} />
+      <div
+        className="bg-secondary absolute z-0 size-8 rounded-full transition-transform"
+        style={{
+          transform: `translateX(${currentLang === "en" ? 0 : 32}px)`,
+        }}
+      />
     </div>
-  )
+  );
 }
