--- apps/frontend/src/lib/i18n.ts
@@ -1,20 +1,0 @@
-import i18n from "i18next";
-import { initReactI18next } from "react-i18next";
-import Backend from "i18next-fs-backend";
-import LanguageDetector from "i18next-browser-languagedetector";
-
-i18n
-  .use(initReactI18next)
-  .use(LanguageDetector)
-  .use(Backend)
-  .init({
-    fallbackLng: "ja",
-    backend: {
-      loadPath: "locales/{{lng}}/{{ns}}.json",
-    },
-    interpolation: {
-      escapeValue: false,
-    },
-  });
-
-export default i18n;
