--- apps/frontend/src/locales/en/translation.json
@@ -1,3 +1,0 @@
-{
-  "welcome": "Welcome!"
-}
