--- apps/frontend/src/locales/ja/translation.json
@@ -1,3 +1,0 @@
-{
-  "welcome": "ようこそ！"
-}
