+++ apps/frontend/src/router.tsx
@@ -1,8 +1,8 @@
-import { createRouter as createTanStackRouter } from '@tanstack/react-router'
-import { routeTree } from './routeTree.gen'
-import { QueryClient } from '@tanstack/react-query'
-import { routerWithQueryClient } from '@tanstack/react-router-with-query'
-
+import { createRouter as createTanStackRouter } from "@tanstack/react-router";
+import { routeTree } from "./routeTree.gen";
+import { QueryClient } from "@tanstack/react-query";
+import { routerWithQueryClient } from "@tanstack/react-router-with-query";
+import { i18n } from "./serverFunctions/i18n-config";
 
 export function createRouter() {
   const queryClient = new QueryClient();
@@ -10,17 +10,20 @@ export function createRouter() {
   return routerWithQueryClient(
     createTanStackRouter({
       routeTree,
-      context: { queryClient },
+      context: {
+        queryClient,
+        crumb: "",
+        lang: null,
+      },
       defaultPreload: "intent",
       scrollRestoration: true,
     }),
     queryClient
-  )
-
+  );
 }
 
-declare module '@tanstack/react-router' {
+declare module "@tanstack/react-router" {
   interface Register {
-    router: ReturnType<typeof createRouter>
+    router: ReturnType<typeof createRouter>;
   }
 }
