+++ apps/frontend/src/routes/$lang/_layout/data-provision/index.tsx
@@ -1,0 +1,9 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/$lang/_layout/data-provision/")({
+  component: RouteComponent,
+});
+
+function RouteComponent() {
+  return <div>Hello "/data-provision/"!</div>;
+}
