+++ apps/frontend/src/routes/$lang/_layout/data-provision/navigation/index.tsx
@@ -1,0 +1,10 @@
+import { Card } from "@/components/Card";
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute(
+  "/$lang/_layout/data-provision/navigation/"
+)({ component: RouteComponent });
+
+function RouteComponent() {
+  return <Card caption="申請システムの前に">Content goes here</Card>;
+}
