--- apps/frontend/src/routes/(other)/_layout/about-data/index.tsx
@@ -1,12 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-import { Markdown } from "@/components/Markdown";
-// import aboutContent from "@/content/about-content.md";
-
-export const Route = createFileRoute("/(other)/_layout/about-data/")({
-  component: About,
-});
-
-function About() {
-  return <Markdown markdown={`hello *hello*`} />;
-}
