--- apps/frontend/src/routes/(other)/_layout/achievements/index.tsx
@@ -1,10 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router"
-
-export const Route = createFileRoute("/(other)/_layout/achievements/")({
-  component: RouteComponent,
-  pendingComponent: () => <div>Loading...</div>,
-})
-
-function RouteComponent() {
-  return <div>Hello "/results/"!</div>
-}
