--- apps/frontend/src/routes/(other)/_layout/data-provision/index.tsx
@@ -1,9 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/(other)/_layout/data-provision/")({
-  component: RouteComponent,
-});
-
-function RouteComponent() {
-  return <div>Hello "/data-provision/"!</div>;
-}
