--- apps/frontend/src/routes/(other)/_layout/data-provision/navigation/index.tsx
@@ -1,10 +1,0 @@
-import { Card } from "@/components/Card";
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute(
-  "/(other)/_layout/data-provision/navigation/"
-)({ component: RouteComponent });
-
-function RouteComponent() {
-  return <Card caption="申請システムの前に">Content goes here</Card>;
-}
