+++ apps/frontend/src/serverFunctions/locale.ts
@@ -1,0 +1,29 @@
+import { createServerFn } from "@tanstack/react-start";
+import { getWebRequest } from "@tanstack/react-start/server";
+import { match as matchLocale } from "@formatjs/intl-localematcher";
+import Negotiator from "negotiator";
+import { i18n, Locale } from "./i18n-config";
+
+function getLocale(request: Request): Locale {
+  // Negotiator expects plain object so we need to transform headers
+  const negotiatorHeaders: Record<string, string> = {};
+  request.headers.forEach((value, key) => (negotiatorHeaders[key] = value));
+
+  // @ts-expect-error locales are readonly
+  const locales: string[] = i18n.locales;
+
+  // Use negotiator and intl-localematcher to get best locale
+  const languages = new Negotiator({ headers: negotiatorHeaders }).languages(
+    locales
+  );
+
+  const locale = matchLocale(languages, locales, i18n.defaultLocale) as Locale;
+
+  return locale;
+}
+
+export const getLocaleFn = createServerFn().handler(async (ctx) => {
+  const request = getWebRequest();
+
+  return getLocale(request);
+});
