+++ apps/frontend/vite.config.ts
@@ -1,6 +1,7 @@
 import { tanstackStart } from "@tanstack/react-start/plugin/vite";
 import { defineConfig } from "vite";
 import tsConfigPaths from "vite-tsconfig-paths";
+import { paraglideVitePlugin as paraglide } from "@inlang/paraglide-js";
 
 export default defineConfig({
   server: {
@@ -10,6 +11,13 @@ export default defineConfig({
   plugins: [
     tsConfigPaths({
       projects: ["./tsconfig.json"],
+    }),
+    paraglide({
+      project: "./project.inlang",
+      outdir: "./src/paraglide",
+      outputStructure: "message-modules",
+      cookieName: "PARAGLIDE_LOCALE",
+      strategy: ["cookie", "url", "baseLocale"],
     }),
     tanstackStart({
       public: {
